feebugs

+++ app/src/main/java/com/example/fee_management_new/Api/AllTransactionList.java
@@ -35,8 +35,8 @@ public class AllTransactionList {
     public boolean gstApplicable;
     public String discount_details;
     public Object addition_details;
-    public int total_discount;
-    public int total_addition;
+    public float total_discount;
+    public float total_addition;
     public String payment_type;
     public String cheque_no;
     public User user;
@@ -170,11 +170,11 @@ public class AllTransactionList {
         return addition_details;
     }
 
-    public int getTotal_discount() {
+    public float getTotal_discount() {
         return total_discount;
     }
 
-    public int getTotal_addition() {
+    public float getTotal_addition() {
         return total_addition;
     }
